reimplemented report runner with argparse
